Clean up various compiler warnings that have crept into the code.  This by no means get's them all, but it's a start.

+++ src/MultiPlayer/multiplayrxmgr.cxx
@@ -233,7 +233,7 @@ void FGMultiplayRxMgr::ProcessData(void) {
 
             // Data received
             if (iBytes > 0) {
-                if (iBytes >= sizeof(MsgHdr)) {
+                if (iBytes >= (int)sizeof(MsgHdr)) {
 
                     // Read header
                     MsgHdr = (T_MsgHdr *)sMsg;
